+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] (in 'dev')
 
-### Fixed
+### Changed
 
 - Reset .cache folder owner every time the container is started
+- Use the SIDs for users/groups from Azure instead of a "randomly" generated one.
+However, you can enable the old handling by setting the env var `LDAP_SAMBA_USEAZURESID` to `false`.
+
 
 ## [1.7.0] - 2022-03-19